projectile interactions complete!
