+++ Assets/Entities/Player/Weapons/StarterBow/Basic Arrow.prefab
@@ -12,10 +12,11 @@ GameObject:
   - component: {fileID: 6466724955343105945}
   - component: {fileID: 6466724955343105946}
   - component: {fileID: 7683933309613658657}
+  - component: {fileID: 7204990223835146144}
   - component: {fileID: 3947936836091602537}
-  - component: {fileID: 6534887806168602107}
+  - component: {fileID: 7725175732944981401}
   - component: {fileID: 6466724955343105951}
-  m_Layer: 0
+  m_Layer: 11
   m_Name: Basic Arrow
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -102,6 +103,22 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
+--- !u!70 &7204990223835146144
+CapsuleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6466724955343105947}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_Size: {x: 0.5, y: 1.5}
+  m_Direction: 0
 --- !u!114 &3947936836091602537
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -114,7 +131,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bfc94850b0bbeaf4993f4472fc6ae746, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &6534887806168602107
+--- !u!114 &7725175732944981401
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -123,13 +140,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 6466724955343105947}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0682da8876652e44e8cb34f04e479023, type: 3}
+  m_Script: {fileID: 11500000, guid: daecfb3790d307d4394a237ad17a1b59, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Speed: 82
-  Gravity: 1
-  GravityMax: 20
-  DragX: 0.5
+  XSpeed: 82
+  YSpeed: 0
 --- !u!114 &6466724955343105951
 MonoBehaviour:
   m_ObjectHideFlags: 0
