+++ Assets/Scenes/Combat.unity
@@ -955,7 +955,7 @@ PrefabInstance:
     - target: {fileID: 3094655183015339571, guid: 774dc8e8a2dbf5e45b005154c655268e,
         type: 3}
       propertyPath: m_IsActive
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 3094655183015339575, guid: 774dc8e8a2dbf5e45b005154c655268e,
         type: 3}
