+++ Assets/Scripts/Gameplay/Motion Controllers/CharacterMotionController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: daecfb3790d307d4394a237ad17a1b59
+guid: 0682da8876652e44e8cb34f04e479023
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
